refactor: 약 entity 수정 #48
Medicine.java- setter 추가
Ingredient.java - 정적 팩터리 메서드 추가
MedicineIngredient.java - quantity 타입 int → double 수정
Unit.java - 단위 한글로 수정

+++ BE/ytt/src/main/java/com/example/ytt/domain/medicine/domain/MedicineIngredient.java
@@ -25,13 +25,13 @@ public class MedicineIngredient {
     private Ingredient ingredient;
 
     @Column(name = "quantity", nullable = false)
-    private int quantity;
+    private double quantity;
 
     @Enumerated(EnumType.STRING)
     private Unit unit;
 
     @Builder
-    public MedicineIngredient(final Medicine medicine, final Ingredient ingredient, final int quantity, final Unit unit) {
+    public MedicineIngredient(final Medicine medicine, final Ingredient ingredient, final double quantity, final Unit unit) {
         Assert.notNull(medicine, "약은 필수입니다.");
         Assert.notNull(ingredient, "성분은 필수입니다.");
         Assert.isTrue(quantity > 0, "수량은 0보다 커야합니다.");
@@ -43,7 +43,7 @@ public class MedicineIngredient {
         this.unit = unit;
     }
 
-    public static MedicineIngredient of(final Medicine medicine, final Ingredient ingredient, final int quantity, final Unit unit) {
+    public static MedicineIngredient of(final Medicine medicine, final Ingredient ingredient, final double quantity, final Unit unit) {
         return MedicineIngredient.builder()
                 .medicine(medicine)
                 .ingredient(ingredient)
@@ -52,7 +52,7 @@ public class MedicineIngredient {
                 .build();
     }
 
-    public static MedicineIngredient of(final Medicine medicine, final Ingredient ingredient, final int quantity, final String unit) {
+    public static MedicineIngredient of(final Medicine medicine, final Ingredient ingredient, final double quantity, final String unit) {
         return of(medicine, ingredient, quantity, Unit.from(unit));
     }
 }
